Reviewed notification side and common infrastructure
