+++ app/src/main/java/com/tari/android/wallet/notification/NotificationBroadcastReceiver.kt
@@ -38,8 +38,8 @@ import android.content.Intent
 import com.orhanobut.logger.Logger
 import com.tari.android.wallet.event.EventBus
 import com.tari.android.wallet.model.recovery.WalletRestorationResult
-import com.tari.android.wallet.ui.activity.SplashActivity
-import com.tari.android.wallet.ui.activity.home.HomeActivity
+import com.tari.android.wallet.ui.fragment.splash.SplashActivity
+import com.tari.android.wallet.ui.fragment.home.HomeActivity
 import com.tari.android.wallet.ui.fragment.restore.restore.WalletRestoreActivity
 
 class NotificationBroadcastReceiver : BroadcastReceiver() {
