+++ app/src/main/java/com/tari/android/wallet/notification/TxCanceledViewHolder.kt
@@ -30,7 +30,7 @@
  * OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.tari.android.wallet.ui.notification
+package com.tari.android.wallet.notification
 
 import android.app.KeyguardManager
 import android.content.Context
