+++ app/src/main/java/com/tari/android/wallet/ui/fragment/home/HomeViewModel.kt
@@ -1,4 +1,4 @@
-package com.tari.android.wallet.ui.activity.home
+package com.tari.android.wallet.ui.fragment.home
 
 import com.tari.android.wallet.ui.common.CommonViewModel
