+++ app/src/main/java/com/tari/android/wallet/ui/fragment/splash/SplashActivity.kt
@@ -30,7 +30,7 @@
  * OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.tari.android.wallet.ui.activity
+package com.tari.android.wallet.ui.fragment.splash
 
 import android.app.Activity
 import android.content.Intent
